pyproject.toml
[tool.poetry]
name = "hub"
version = "0.1.0"
description = "Integration Hub - Phase 0 foundations"
authors = ["Abdelrahman Mahmoud"]
readme = "README.md"
packages = [{ include = "app" }, { include = "worker" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
sqlalchemy = "^2.0.45"
asyncpg = "^0.31.0"
alembic = "^1.18.0"
python-json-logger = "^4.0.0"
httpx = "^0.28.1"
redis = "^7.1.0"
celery = "^5.6.2"
kombu = "^5.6.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "^46.0.3"
opentelemetry-api = "^1.39.1"
opentelemetry-sdk = "^1.39.1"
opentelemetry-exporter-otlp = "^1.39.1"
opentelemetry-instrumentation-fastapi = "^0.60b1"
opentelemetry-instrumentation-sqlalchemy = "^0.60b1"
opentelemetry-instrumentation-requests = "^0.60b1"
opentelemetry-instrumentation-logging = "^0.60b1"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
ruff = "^0.14.11"
mypy = "^1.19.1"
python-dotenv = "^1.2.1"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
